Camera Class, now with methods to get rotation offset and distance.
The rotation offset method will return the number of pixels to move the
target to the center of the camera. For example it may return -27
pixels further testing required to see if the robot should rotate left
or right based on negative or positive.

 The distance method will return the distance to the target in feet, for 
 example, it has returned 3.8 feet before. Further testing required to
see if the measurements are accurate.
 
 Other changes have been made to the particle tests, I have removed the
method call invoking the test for "convexhullarea" because it returned
0.03 on a standard 0-100 scale for all tests. This means the SCORE_MIN
would have to be ridiculously low for the test to have passed, the 
 other two tests return values greater than 60 for comparison.