I optimized the getTarget method:
I changed the loop that iterates through the vector list of particles,
it used to set all of the values for each particle report. Now it only
sets the area and index value of the particle report, which are
necessary for sorting. The rest of the values are only set for the
largest particle which is found through sorting.